@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 
 class DownloadRequest {
   @required
-  final String url;
+  final String? url;
 
   DownloadRequest({this.url});
 }
